requirements-bootstrap.txt
PyYAML>=6.0
openai>=1.0.0
litellm>=1.75.5
